t0500: New regression test for git add of a path that contains a .git directory
This test was created in 2011 to address a regression that was introduced with
18e051a3981f38db08521bb61ccf7e4571335353 that broke Xcode's snapshotting feature.

The issue was eventually addressed with ddc2a6281595fd24ea01497c496f88c40a59562f
in 2014, so all that remains is this test for the original issue.

This test script mimics the behavior of the Xcode snapshotting feature that
triggered the original problem.

TODO:
    Should this be rolled into another test script?  Should it be renamed?

Signed-off-by: Jeremy Huddleston Sequoia <jeremyhu@apple.com>